Refactor MainsScreen and MainsVerdictOverlay to streamline image upload process and enhance navigation flow
